Moved the auto code to the right folder.
