+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/BlueGoldAutoOpMode.java
@@ -1,0 +1,12 @@
+package org.firstinspires.ftc.teamcode.autonomous;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+@Autonomous(name="Blue Gold", group="Autonomous")
+public class BlueGoldAutoOpMode extends BaseAutoOpMode {
+
+    public void runOpMode(){
+
+        InitRobot();
+    }
+}
