--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/BlueGoldAutoOpMode.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.teleop;
-
-public class BlueGoldAutoOpMode extends BaseAutoOpMode {
-
-    public void runOpMode(){
-
-        InitRobot();
-    }
-}
